+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Current
 
+* Control I/O on Windows works without many bugs using legacy API, and mostly wor
+
+# 0.2.0
+
 * Added MIDI input and output for Windows.
 * Preliminary work on control I/O has begun.
 * Audio testcase is now properly named.